feat(comment): 重构评论系统支持异步编排和虚拟线程
- 新增 CommentIndexDto 数据传输对象
- 添加虚拟线程执行器配置支持
- 实现评论列表异步编排组装功能
- 优化评论查询使用多级缓存策略
- 添加评论内容和点赞信息批量查询
- 重构评论点赞数缓存重建逻辑
- 新增 LikeInfoVo 点赞信息承载类
- 优化数据库索引提升查询性能
- 修复评论内容缓存键名格式问题
- 移除用户类型字段简化数据结构

+++ src/main/java/com/losgai/sys/mapper/LikeMapper.java
@@ -40,7 +40,7 @@ public interface LikeMapper {
 
     Long existByUserIdAndCommentId(Long userId, Long commentId);
 
-    Long countByCommentId(Long commentId);
+    Integer countByCommentId(Long commentId);
 
     List<Like> selectLatestLikes(Long userId, Integer likeLimit);
 }
